add a compact version of audiogram chart without symbols
